EXR and non Linear format are now exactly the same !!
The last retake was about ACES and Non linear format, the problem was the little difference in vibrance and clarity.

Now the two format are the same, we just need to set the non linear format to RAW in the compositing software such as Nuke or Natron, After Effects etc etc...

Basically I need to set the project as linear/ACEScg, so that every linear files will be interpreted as ACEScg linear colorspace, and every non linear entry need to be set as RAW input, so that no ACEScg filters are applyed, normally we need to set up ACEScg/sRGB entry, but RAW is the same thing in my setup.

also added transform and rotation instances